+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fixed an issue where effects were not re-enabled when switching from an Airplay device back to phone output (#678)
 - When connected to CarPlay the Up Next Queue will more consistently display at the top of the podcasts list (#680)
 - Fixed an issue where the Up Next queue doesn't continue playing the next episode when connected to AirPlay (#676)
+- Fixed an issue where the swipe actions in Up Next could be triggered when trying to reorder a row (#684)
 
 
 7.31